+++ Engine/Graphics/Primitives/circle.py
@@ -48,10 +48,10 @@ class Circle(Base):
         
         # Optimization
         
-        if self.position.x+self.size.x < 0 or self.position.y+self.size.y < 0:
-            return
-
         if self.window:
+            if self.position.x+self.size.x < 0 or self.position.y+self.size.y < 0:
+                return
+
             if self.position.x-self.size.x > self.window.current_window_sizes[0] or self.position.y-self.size.y > self.window.current_window_sizes[1]:
                 return